Fix navigation logging
not sure exactly what was wrong here; I think it was related to the ref

also got rid of a hook because this did not need to be a hook
