+++ js/navigation-ref.js
@@ -1,12 +1,11 @@
-import React from 'react';
-
-export const navigationRef = React.createRef();
+let _navigationRef = null;
+export const setNavigationRef = (ref) => _navigationRef = ref;
 
 export const navigate = (route, params) => {
-  navigationRef.current?.navigate(route, params);
+  _navigationRef?.navigate(route, params);
 };
 
 export const pop = () => {
   // gotta call goBack like a dummy since pop only exists on stacknavigator and this is a level up
-  navigationRef.current?.goBack();
+  _navigationRef?.goBack();
 };
